demos working

+++ openssl/build.gradle.kts
@@ -20,35 +20,38 @@ plugins {
 
 version = "OpenSSL_1_1_1q"
 
+
+val KonanTarget.openSSLNotBuilt: Boolean
+  get() = !opensslPrefix(project).resolve("include/openssl/ssl.h").exists()
+
+
 val opensslGitDir = rootProject.file("repos/openssl")
 
+tasks.register("generateCInteropsDef") {
+  inputs.file("src/openssl_header.def")
+  outputs.file("src/openssl.def")
+  doFirst {
+    val outputFile = outputs.files.files.first()
+    println("Generating $outputFile")
 
- tasks.register("generateCInteropsDef") {
-   inputs.file("src/openssl_header.def")
-   outputs.file("src/openssl.def")
-   doFirst {
-     val outputFile = outputs.files.files.first()
-     println("Generating $outputFile")
-
-     outputFile.printWriter().use { output ->
-       output.println(inputs.files.files.first().readText())
-       kotlin.targets.withType<KotlinNativeTarget>().forEach {
-         val konanTarget = it.konanTarget
-         output.println("compilerOpts.${konanTarget.name} = -Ibuild/kotlinxtras/openssl/${konanTarget.platformName}/include \\")
-         output.println("\t-I/usr/local/kotlinxtras/libs/openssl/${konanTarget.platformName}/include ")
-         output.println("linkerOpts.${konanTarget.name} = -Lbuild/kotlinxtras/openssl/${konanTarget.platformName}/lib \\")
-         output.println("\t-L/usr/local/kotlinxtras/libs/openssl/${konanTarget.platformName}/lib ")
-       }
-     }
-   }
- }
+    outputFile.printWriter().use { output ->
+      output.println(inputs.files.files.first().readText())
+      kotlin.targets.withType<KotlinNativeTarget>().forEach {
+        val konanTarget = it.konanTarget
+        output.println("compilerOpts.${konanTarget.name} = -Ibuild/kotlinxtras/openssl/${konanTarget.platformName}/include \\")
+        output.println("\t-I/usr/local/kotlinxtras/libs/openssl/${konanTarget.platformName}/include ")
+        output.println("linkerOpts.${konanTarget.name} = -Lbuild/kotlinxtras/openssl/${konanTarget.platformName}/lib \\")
+        output.println("\t-L/usr/local/kotlinxtras/libs/openssl/${konanTarget.platformName}/lib ")
+      }
+    }
+  }
+}
 
 fun srcPrepare(target: KonanTarget) =
   tasks.register("srcPrepare${target.platformName.capitalize()}", Exec::class) {
     val srcDir = target.opensslSrcDir(project)
-    onlyIf {
-      !srcDir.exists()
-    }
+    outputs.dir(srcDir)
+    onlyIf { target.openSSLNotBuilt }
     commandLine(BuildEnvironment.gitBinary, "clone", opensslGitDir, srcDir)
   }
 
@@ -59,7 +62,9 @@ fun configureTask(target: KonanTarget) =
     val srcDir = target.opensslSrcDir(project)
     workingDir(srcDir)
     environment(target.buildEnvironment())
-    outputs.file(srcDir.resolve("Makefile"))
+    val makeFile = srcDir.resolve("Makefile")
+    onlyIf { target.openSSLNotBuilt }
+    outputs.file(makeFile)
     doFirst {
       println("OpenSSL Configure $target ..")
     }
@@ -81,19 +86,24 @@ fun buildTask(target: KonanTarget) =
   tasks.register("build${target.platformName.capitalized()}", Exec::class) {
     dependsOn("srcConfigure${target.platformName.capitalized()}")
 
+
     val installDir = target.opensslPrefix(project)
 
     dependsOn(target.konanDepsTask(project))
 
     workingDir(target.opensslSrcDir(project))
 
-    outputs.file(installDir.resolve("include/openssl/ssl.h"))
+    val outputFile = installDir.resolve("include/openssl/ssl.h")
+    outputs.file(outputFile)
+    onlyIf { target.openSSLNotBuilt }
+
     environment(target.buildEnvironment())
     group = BasePlugin.BUILD_GROUP
     commandLine("make", "install_sw")
-    /*doLast {
+    doLast {
+      if (didWork)
       target.opensslSrcDir(project).deleteRecursively()
-    }*/
+    }
   }
